smell fix

+++ trunk/po0_220227/task_03/src/Real.cpp
@@ -1,7 +1,7 @@
 #include "Real.h"
 #include <iostream>
 
-Real::Real(const int _value) : value(_value)
+Real::Real(const float _value) : value(_value)
 {
 }
